fixed type for divide method

+++ first.task/src/lab2/Calculator.java
@@ -13,10 +13,10 @@ public class Calculator {
     public static double divide(double a, double b) {
         return a / b;
     }
-    public static long divide(long a, long b) {
+    public static double divide(long a, long b) {
         return a / b;
     }
-    public static int divide(int a, int b) {
+    public static double divide(int a, int b) {
         return a / b;
     }
     public static double multiply(double a, double b) {
